c2rust-transpile: add partial bfloat16 support
like Half, this is basically just a stub for now

supporting these types properly could use the `half` crate, but this is just to get basic AArch64 translation working

this type was also added in Clang/LLVM 11, so ifdef'd accordingly

we need to handle these now because ARMv8.6-a adds the BFloat16/__bf16 type as a vector element of one of the SVE vector types just added

+++ c2rust-transpile/src/convert_type.rs
@@ -332,7 +332,7 @@ impl TypeConverter {
             CTypeKind::LongDouble => Ok(mk().path_ty(mk().path(vec!["f128", "f128"]))),
             CTypeKind::Float => Ok(mk().path_ty(mk().path(vec!["libc", "c_float"]))),
             CTypeKind::Int128 => Ok(mk().path_ty(mk().path(vec!["i128"]))),
-            CTypeKind::UInt128 => Ok(mk().path_ty(mk().path(vec!["u128"]))),
+            CTypeKind::UInt128 => Ok(mk().path_ty(mk().path(vec!["bf16"]))),
 
             CTypeKind::Pointer(qtype) => self.convert_pointer(ctxt, qtype),
